e2e: split runtime_class_test.go into functional, burst, and benchmark files
Split the 1330-line monolith into four focused files:
- runtime_class_test.go (240 lines) — functional CI-suitable tests
- runtime_class_helpers_test.go (139 lines) — shared helpers
- runtime_class_burst_test.go (792 lines) — stress/longevity test
- runtime_class_bench_test.go (246 lines) — Go benchmarks

No behavior changes — pure file reorganization.
